cpsv: support for SAF SaTimeT timeout variable in MDS transport [#1658] Issue : In current implementation their is a mismatch of variable sizes between SAF API accepting SaTimeT variable type for timeout ( 64 bit) and MDS transport implementation accepting (32 bit) timeout variable , the MDS transport should adopt the size of the timeout parameter from 32 bits to 64 bits.
Fix :
Now MDS provide interface to accept SaTimeT type timeout variable from SAF APIS

+++ osaf/services/saf/cpsv/cpnd/cpnd_tmr.c
@@ -130,11 +130,11 @@ uint32_t cpnd_tmr_start(CPND_TMR *tmr, SaTimeT duration)
 	}
 
 	if (tmr->is_active == false) {
-		m_NCS_TMR_START(tmr->tmr_id, (uint32_t)duration, cpnd_timer_expiry, (void *)tmr);
+		m_NCS_TMR_START(tmr->tmr_id, duration, cpnd_timer_expiry, (void *)tmr);
 		tmr->is_active = true;
 	} else if (tmr->is_active == true) {
 		m_NCS_TMR_STOP(tmr->tmr_id);
-		m_NCS_TMR_START(tmr->tmr_id, (uint32_t)duration, cpnd_timer_expiry, (void *)tmr);
+		m_NCS_TMR_START(tmr->tmr_id, duration, cpnd_timer_expiry, (void *)tmr);
 	}
 
 	return (NCSCC_RC_SUCCESS);
